+++ src/controllers/SettingsController.php
@@ -1,0 +1,72 @@
+<?php
+/**
+ * Fathom Analytics plugin for Craft CMS 3.x
+ *
+ * Statistics and chart widgets for Fathom analytics.
+ *
+ * @link      https://stenvdb.be
+ * @copyright Copyright (c) 2020 Sten Van den Bergh
+ */
+
+namespace stenvdb\fathomanalytics\controllers;
+
+use stenvdb\fathomanalytics\FathomAnalytics;
+
+use Craft;
+use craft\web\Controller;
+
+/**
+ * Settings Controller
+ *
+ * Generally speaking, controllers are the middlemen between the front end of
+ * the CP/website and your plugin’s services. They contain action methods which
+ * handle individual tasks.
+ *
+ * A common pattern used throughout Craft involves a controller action gathering
+ * post data, saving it on a model, passing the model off to a service, and then
+ * responding to the request appropriately depending on the service method’s response.
+ *
+ * Action methods begin with the prefix “action”, followed by a description of what
+ * the method does (for example, actionSaveIngredient()).
+ *
+ * https://craftcms.com/docs/plugins/controllers
+ *
+ * @author    <%- pluginAuthorName %>
+ * @package   <%= pluginHandle %>
+ * @since     <%= pluginVersion %>
+ */
+
+class SettingsController extends Controller
+{
+
+    // Protected Properties
+    // =========================================================================
+
+    protected $allowAnonymous = [];
+
+    // Public Methods
+    // =========================================================================
+
+    public function actionSave()
+    {
+        $this->requireCpRequest();
+
+        $settings = Craft::$app->getRequest()->getBodyParam('settings', []);
+        $plugin = Craft::$app->getPlugins()->getPlugin('fathom-analytics');
+
+        if (!Craft::$app->getPlugins()->savePluginSettings($plugin, $settings)) {
+            Craft::$app->getSession()->setError(Craft::t('app', "Couldn't save plugin settings."));
+
+            // Send the plugin back to the template
+            Craft::$app->getUrlManager()->setRouteParams([
+                'plugin' => 'fathom-analytics',
+            ]);
+
+            return null;
+        }
+
+        Craft::$app->getSession()->setNotice(Craft::t('app', 'Plugin settings saved.'));
+
+        return $this->redirectToPostedUrl();
+    }
+}