+++ src/models/Settings.php
@@ -48,6 +48,20 @@ class Settings extends Model
 
     public $trackingId;
 
+    public $injectTracking = false;
+
+    public $trackingCode = "(function(f, a, t, h, o, m){
+    a[h]=a[h]||function(){
+      (a[h].q=a[h].q||[]).push(arguments)
+    };
+    o=f.createElement(\'script\'),
+      m=f.getElementsByTagName(\'script\')[0];
+    o.async=1; o.src=t; o.id=\'fathom-script\';
+    m.parentNode.insertBefore(o,m)
+  })(document, window, \'//{{ baseUri }}/tracker.js\', \'fathom\');
+  fathom(\'set\', \'siteId\', \'{{ trackingId }}\');
+  fathom(\'trackPageview\');";
+
     // Public Methods
     // =========================================================================