+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2020-05-18
+### Added
+- Editable tracking code snippet in plugin settings
+- This plugin can nog automatically inject the tracking snippet into your site (disabled by default)
+
 ## 1.0.1 - 2020-04-10
 ### Fixed
 - Case sensitive namespace fixed